requirements.txt
numpy==2.0.1
pandas==2.2.2
pyyaml==6.0.2
click==8.1.7
